requirements.txt
streamlit
pandas
requests
fpdf
google-generativeai